Make Mod Menu an optional dependency
The HUD editor is reachable via its keybinding and hosts the Settings
button itself, so Mod Menu only contributes a configure button in its
mod list. The Cloth Config screen builder moves out of the ModMenu
entrypoint class into SettingsScreenBuilder so the editor never touches
ModMenuApi, which only classloads when Mod Menu is installed. modmenu
moves from depends to suggests in fabric.mod.json.

Verified with the 26.2 launch test both with Mod Menu loaded and with it
removed from the runtime mods.
